+++ spec/controllers/admin/customers_controller_spec.rb
@@ -11,7 +11,7 @@ describe Admin::CustomersController, type: :controller do
 
     it "passes the search term to the view" do
       locals = capture_view_locals do
-        get :index, kwarg_params(search: "foo")
+        get :index, search: "foo"
       end
 
       expect(locals[:search_term]).to eq("foo")
@@ -21,6 +21,13 @@ describe Admin::CustomersController, type: :controller do
       locals = capture_view_locals { get :index }
 
       expect(locals[:page]).to be_instance_of(Administrate::Page::Collection)
+    end
+
+    it "shows the search bar" do
+      customer = create(:customer)
+
+      locals = capture_view_locals { get :index }
+      expect(locals[:show_search_bar]).to be_truthy
     end
   end
 
@@ -173,27 +180,6 @@ describe Admin::CustomersController, type: :controller do
       delete :destroy, kwarg_params(id: customer.to_param)
 
       expect(response).to redirect_to(admin_customers_url)
-    end
-  end
-
-  def capture_view_locals
-    allow(@controller).to receive(:render)
-    yield
-
-    locals = nil
-    expect(@controller).to have_received(:render).at_least(1).times do |*args|
-      args.each do |arg|
-        locals ||= arg.try(:fetch, :locals, nil)
-      end
-    end
-    locals
-  end
-
-  def kwarg_params(args)
-    if Rails::VERSION::MAJOR >= 5
-      { params: args }
-    else
-      args
     end
   end
 end